requirements.txt
streamlit
spacy
scikit-learn
pandas
matplotlib
pdfminer.six
python-docx









































